Integrating from main to FRANKLIN

Change 83902 by jhahm@jhahm-mac on 2008/02/14 15:19:07

	bug: 22645

	Incoming directory sweeper thread should handle race condition with 
	message delivery thread.  Sweeper can call listFiles() on the 
	incoming directory while delivery thread is in the middle of 
	saving-then-moving a file to incoming directory.  Sweeper can 
	encounter files that were there when listFiles() was called but gone 
	when it calls File.lastModified() to do age check.  Detect this case 
	by checking for return value of 0L from lastModified().

	The file can exist at the time of lastModified() call but may get 
	moved out just before sweeper calls File.delete().  delete() will 
	return false in that case.  Sweeper should check for file existence 
	before logging warning about delete failure.  This should be a very 
	rare case, because files that were older than sweeper age threshold 
	aren't likely to be moved by any other thread.

http://bugzilla.zimbra.com/show_bug.cgi?id=22645

Copied from Perforce
 Change: 83903

+++ src/java/com/zimbra/cs/store/FileBlobStore.java
@@ -36,8 +36,6 @@ import java.util.List;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
 
-import javax.mail.util.SharedByteArrayInputStream;
-
 import com.zimbra.common.localconfig.LC;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.util.ByteUtil;
@@ -532,17 +530,24 @@ public class FileBlobStore extends StoreManager {
 
         				File file = files[i];
                 		if (file.isDirectory()) continue;
-                		long age = startTime - file.lastModified();
-                		if (age >= mMaxAgeMS) {
-                			boolean deleted = file.delete();
-                			if (!deleted) {
-                				mLog.warn("Sweeper unable to delete " +
-                						  file.getAbsolutePath());
-                			} else if (mLog.isDebugEnabled()) {
-                				mLog.debug("Sweeper deleted " +
-                						   file.getAbsolutePath());
-                				numDeleted++;
-                			}
+                		long lastMod = file.lastModified();
+                		// lastModified() returns 0L if file doesn't exist (i.e. deleted by another thread
+                		// after this thread did directory.listFiles())
+                		if (lastMod > 0L) {
+                    		long age = startTime - lastMod;
+                    		if (age >= mMaxAgeMS) {
+                    			boolean deleted = file.delete();
+                    			if (!deleted) {
+                    			    // Let's warn only if delete failure wasn't caused by file having been
+                    			    // by someone else already.
+                    			    if (file.exists())
+                    			        mLog.warn("Sweeper unable to delete " + file.getAbsolutePath());
+                    			} else if (mLog.isDebugEnabled()) {
+                    				mLog.debug("Sweeper deleted " +
+                    						   file.getAbsolutePath());
+                    				numDeleted++;
+                    			}
+                    		}
                 		}
                 	}
                 	synchronized (this) {
